+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Changed
 
+- concord-server: fix empty secret name in project repository
+entries returned by the API;
 - concord-server: added environment variable to override the
 server's password.