+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Added
 
+- concord-console: ability to start a new process from the project's
+form;
 - ansible: support for saving and using Ansible's "limit" files;
 - concord-server, concord-console: support for boolean form fields;
 - ansible: Inventory lookup plugin.